fixing condensation
