+++ particula/tests/particle_test.py
@@ -279,3 +279,39 @@ def test_coulomb_enh():
 #     assert negative_ion.dimensioned_coagulation_kernel(
 #         positive_particle,
 #     ).check("[length]**3/[time]")
+
+
+def test_condensation_stuff():
+    """ testing condensation stuff
+    """
+
+    simple_cond_kwargs = {
+    "mode": 200e-9,  # 200 nm median
+    "nbins": 1000,  # 1000 bins
+    "nparticles": 2e6,  # 1e4 #
+    "volume": 1e-6,  # per 1e-6 m^3 (or 1 cc)
+    "gsigma": 1.25,  # relatively narrow
+    "cutoff": .99999,  # let's take it all lol
+    "vapor_radius": 1.6e-9,  # 1.6 nm
+    "vapor_density": 1400,  # 1400 kg/m^3
+    "vapor_concentration": 1,  # 1 ug/m^3
+    "vapor_attachment": 1,  # 1
+    "vapor_molec_wt": 200,  # 200 g/mol
+    "something_something": None
+    }
+
+    cond = particle.Particle(**simple_cond_kwargs)
+
+    assert cond.particle_radius.m.shape == (1000,)
+    assert cond.driving_force().u == u.kg/u.m**3
+    # assert cond.driving_force().m.shape == (1,)
+    assert cond.molecular_enhancement().u == u.dimensionless
+    assert cond.molecular_enhancement().m.shape == (1000,)
+    assert cond.condensation_redmass().u == u.kg/u.mol
+    assert cond.condensation_redmass().m.shape == (1000,)
+    assert cond.vapor_speed().u == u.m/u.s
+    assert cond.vapor_speed().m.shape == (1000,)
+    assert cond.vapor_flux().u == u.kg/u.s
+    assert cond.vapor_flux().m.shape == (1000,)
+    assert cond.particle_growth().u == u.m/u.s
+    assert cond.particle_growth().m.shape == (1000,)
